nadfun-sdk-python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nadfun-sdk"
version = "0.1.3"
description = "Python SDK for Nad.fun contracts"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Nadfun", email = "dev@nad.fun"}
]
keywords = ["blockchain", "ethereum", "defi", "trading", "bonding-curve"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "web3>=6.0.0",
    "eth-account>=0.10.0",
    "eth-abi>=4.0.0",
    "eth-utils>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/naddotfun/nadfun-sdk"
Issues = "https://github.com/naddotfun/nadfun-sdk/issues"
Repository = "https://github.com/naddotfun/nadfun-sdk"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
nadfun_sdk = ["abis/*.json"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true